chore: moving transformColors to designSystemUtils
